requirements.txt
# Core dependencies
click==8.1.8
websockets==15.0
colorama==0.4.6

# Database
mysqlclient==2.2.0
mysql-connector-python==8.3.0
SQLAlchemy==2.0.27
alembic==1.13.1
redis==5.0.1
python-dotenv==1.0.0

# Security
cryptography==41.0.7
PyJWT==2.8.0
argon2-cffi==23.1.0

# Terminal UI enhancements
rich==13.7.0
prompt_toolkit==3.0.43

# Testing
pytest==7.4.4
pytest-asyncio==0.23.5